cleaning up a lot of hanging records in extractor
