+++ skyrim-extractor/src/binary-decoders/__tests__/grupHandler.test.ts
