+++ skyrim-extractor/src/binary-decoders/recordParser.ts
@@ -1,6 +1,10 @@
-import { RECORD_HEADER, SUBRECORD_HEADER, OFFSETS } from "./buffer.constants";
+import {
+  RECORD_HEADER,
+  SUBRECORD_HEADER,
+  OFFSETS,
+} from "../utils/buffer.constants";
 import { RecordHeader, SubrecordHeader } from "../types";
-import { debugLog } from "./debugUtils";
+import { debugLog } from "../utils/debugUtils";
 
 /**
  * Low-level record parsing utilities
