+++ skyrim-extractor/src/constants.ts
