+++ skyrim-extractor/src/utils/paths.ts
